remove use of pywsman
pywsman is a set of python bindings built into the openwsman
project. This is a c library which implements an http client with
digest auth and xml processing. As I am not trying to build a generic
wsman layer, it's much simpler to just hard code the message templates
that were built via openwsman, and use requests and eltree for the
http and xml work.

This makes the tool *far* more portable, as it's now pure python. It
also will make the recovery code easier to write as it won't be
handling exceptions passed up from native code.